Improve synth stability, fix FM, and re-enable global effects
- Fixed signed math in FM modulation to handle negative modulator values.
- Increased DMA buffer size and optimized filter coefficients with fast exponential approximation to prevent crashes with higher polyphony.
- Corrected voice allocation logic in noteOn to use the proper voice stealing mechanism.
- Fully implemented serial commands for Granular, Resonator, and Comb effects.
- Added AudioProfiler and unit test suite for diagnostic and verification purposes.
- Expanded Preset system to include enabled flags for all global effects.

+++ Oscillator.cpp
@@ -165,8 +165,9 @@ float Oscillator::process() {
 
 float Oscillator::processWithFM(float fmInput, float fmAmount) {
     // FM synthesis: modulate phase increment
-    uint32_t fmOffset = (uint32_t)(fmInput * fmAmount * basePhaseIncrement_);
-    uint32_t effectiveIncrement = basePhaseIncrement_ + fmOffset;
+    // Use signed math to handle negative fmInput correctly
+    int32_t fmOffset = (int32_t)(fmInput * fmAmount * (float)basePhaseIncrement_);
+    uint32_t effectiveIncrement = (uint32_t)((int32_t)basePhaseIncrement_ + fmOffset);
     
     // Apply pitch mod on top
     if (pitchMod_ != 0.0f) {
